Redirect back to the consumer's request.path_url (without the QUERY_STRING).
This allows us to do a GET to the verify_openid view with the openid provider
as a query string. Before this change, it wouldn't strip the query string off
and would pass in the openid.provider back into the redirect after a successful
login. When this plugin sees an openid provider in the params, it always
assumes it's initializing a login, thus causing a login loop.